Code refinements made
