+++ Pet Reminder/Pet List/Models/PetListViewReference.swift
@@ -1,0 +1,13 @@
+//
+//  PetListViewReference.swift
+//  Pet Reminder
+//
+//  Created by Ege Sucu on 31.08.2023.
+//  Copyright © 2023 Softhion. All rights reserved.
+//
+
+import Foundation
+
+enum PetListViewReference {
+    case petList, settings
+}
